Basic level to walk around setup
The game has a basic level that the player is able to move around and play in, start of the interact and use has been added but nothing to implement with
