+++ Lucky The Crafty Cat/Assets/Scripts/Camera/CameraController.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CameraController : MonoBehaviour
+{
+    // Player character the camera will follow
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
