--- Lucky The Crafty Cat/Assets/Scripts/Movement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fcfffa926769269458b964c592aab7fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
