+++ Lucky The Crafty Cat/Assets/Scripts/Player/Actions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4565ec260b539ac46a32959648c7002c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
