+++ Lucky The Crafty Cat/Library/Bee/artifacts/1900b0aE.dag/Assembly-CSharp.rsp
@@ -352,7 +352,9 @@
 -r:"Library/Bee/artifacts/1900b0aE.dag/UnityEngine.UI.ref.dll"
 -analyzer:"C:/Program Files/Unity/Hub/Editor/2022.3.17f1/Editor/Data/Tools/Unity.SourceGenerators/Unity.Properties.SourceGenerator.dll"
 -analyzer:"C:/Program Files/Unity/Hub/Editor/2022.3.17f1/Editor/Data/Tools/Unity.SourceGenerators/Unity.SourceGenerators.dll"
-"Assets/Scripts/Movement.cs"
+"Assets/Scripts/Camera/CameraController.cs"
+"Assets/Scripts/Player/Actions.cs"
+"Assets/Scripts/Player/Movement.cs"
 -langversion:9.0
 
 /deterministic
