+++ Lucky The Crafty Cat/Library/StateCache/SceneView/8c/8cd7c613bf844de3b80696e27a479d5e.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":0.2752799093723297,"y":-19.178804397583009,"z":13.80478286743164},"rotation":{"x":-0.42412498593330386,"y":0.004343504551798105,"z":-0.0020342273637652399,"w":-0.9055927395820618},"size":27.9813175201416,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":0.631401538848877,"y":-0.20539093017578126,"z":-0.3904228210449219},"rotation":{"x":-0.47609803080558779,"y":-0.02874922566115856,"z":0.015576636418700219,"w":-0.8787873983383179},"size":23.632051467895509,"orthographic":false}
